# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.SafeUser:
    properties:
      creationDate:
        type: string
      id:
        type: integer
      userName:
        type: string
    type: object
  model.Domain:
    properties:
      CreationDate:
        type: string
      CreatorId:
        type: integer
      DomainName:
        type: string
      Id:
        type: integer
    type: object
  model.DomainList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Domain'
        type: array
    type: object
  model.FailureMsg:
    properties:
      error:
        type: string
    type: object
  model.ProposedUser:
    properties:
      CreationDate:
        type: string
      Id:
        type: integer
      Password:
        type: string
      Status:
        type: string
      UserName:
        type: string
    type: object
  model.SuccessMsg:
    properties:
      message:
        type: string
    type: object
  model.User:
    properties:
      CreationDate:
        type: string
      Id:
        type: integer
      PasswordHash:
        type: string
      Status:
        type: string
      UserName:
        type: string
    type: object
  model.UserStatusMsg:
    properties:
      message:
        type: string
      userStatus:
        type: string
    type: object
  model.UsersList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
host: localhost:8000
info:
  contact:
    name: Gary Greene
    url: https://github.com/greeneg/ipmanager
  description: A simple API for managing networks
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: IpManager API
  version: 0.0.6
paths:
  /domain:
    post:
      consumes:
      - application/json
      description: Create a new domain
      parameters:
      - description: Domain data
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/model.Domain'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Create a new domain
      tags:
      - domain
  /domain/{domainname}:
    delete:
      description: Delete a domain
      parameters:
      - description: Domain name
        in: path
        name: domainname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Delete a domain
      tags:
      - domain
  /domain/id/{domainid}:
    get:
      description: Retrieve a domain by Id
      parameters:
      - description: Domain Id
        in: path
        name: domainid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Domain'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      summary: Retrieve a domain by Id
      tags:
      - domain
  /domain/name/{domainname}:
    get:
      description: Retrieve a domain by DomainName
      parameters:
      - description: Domain name
        in: path
        name: domainname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Domain'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      summary: Retrieve a domain by DomainName
      tags:
      - domain
  /domains:
    get:
      description: Retrieve a list of domain
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DomainList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      summary: Retrieve a list of domain
      tags:
      - domain
  /user:
    post:
      consumes:
      - application/json
      description: Add a new user
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ProposedUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Register user
      tags:
      - user
  /user/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      parameters:
      - description: User name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Delete user
      tags:
      - user
  /user/{name}/status:
    get:
      consumes:
      - application/json
      description: Retrieve a user's active status
      parameters:
      - description: User name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserStatusMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a user's active status. Can be either 'enabled' or 'locked'
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Set a user's active status
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      - description: User name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserStatusMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Set a user's active status. Can be either 'enabled' or 'locked'
      tags:
      - user
  /user/id/{id}:
    get:
      description: Retrieve a user by their Id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SafeUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      summary: Retrieve a user by their Id
      tags:
      - user
  /user/name/{name}:
    get:
      description: Retrieve a user by their UserName
      parameters:
      - description: User name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SafeUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      summary: Retrieve a user by their UserName
      tags:
      - user
  /users:
    get:
      description: Retrieve list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UsersList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      summary: Retrieve list of all users
      tags:
      - user
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"